Move storage files to dedicated folder
