+++ src/app/team/team.page.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { HelpService } from '../help/help.service';
-import { ThemesService } from '../settings/themes.service';
+import { ThemesService } from '../settings/settings-storage/themes.service';
 import { TeamService } from './team.service';
 
 @Component({
